backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.5
python-dotenv==0.19.0
pydantic[email]==1.8.2
email-validator==1.1.3
reportlab==4.0.4
alembic==1.12.0